refactor: move ralph utilities to src/oclif-commands/ralph/lib/
Move state.ts, formatter.ts, execution.ts, marker.ts, index.ts from
src/commands/ralph/ to src/oclif-commands/ralph/lib/. Update all
imports in oclif commands to use new paths.
